requirements.txt
Django>=4.2
psycopg2-binary>=2.9
djangorestframework>=3.14
django-cors-headers>=4.0
Pillow>=10.0

